Moves SonarProjectConfig creation to Test Helper
