+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/Hotspots/CookieShouldBeHttpOnlyTest.cs
@@ -22,7 +22,6 @@
 using System.Collections.Generic;
 using Microsoft.CodeAnalysis;
 #endif
-using System.IO;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using SonarAnalyzer.Common;
 using SonarAnalyzer.UnitTest.MetadataReferences;
@@ -34,11 +33,6 @@ namespace SonarAnalyzer.UnitTest.Rules
     [TestClass]
     public class CookieShouldBeHttpOnlyTest
     {
-        private const string ProjectConfigTemplate = @"
-<SonarProjectConfig xmlns=""http://www.sonarsource.com/msbuild/analyzer/2021/1"">
-    <FilesToAnalyzePath>{0}\FilesToAnalyze.txt</FilesToAnalyzePath>
-</SonarProjectConfig>";
-
 #if NETFRAMEWORK // The analyzed code is valid only for .Net Framework
         [TestMethod]
         [TestCategory("Rule")]
@@ -55,7 +49,7 @@ namespace SonarAnalyzer.UnitTest.Rules
             Verifier.VerifyAnalyzer(@"TestCases\WebConfig\CookieShouldBeHttpOnly\HttpOnlyCookiesConfig\CookieShouldBeHttpOnly.cs",
                 new CS.CookieShouldBeHttpOnly(AnalyzerConfiguration.AlwaysEnabled),
                 MetadataReferenceFacade.SystemWeb,
-                CreateSonarProjectConfig(@"TestCases\WebConfig\CookieShouldBeHttpOnly\HttpOnlyCookiesConfig"));
+                TestHelper.CreateSonarProjectConfig(@"TestCases\WebConfig\CookieShouldBeHttpOnly\HttpOnlyCookiesConfig"));
 
         [TestMethod]
         [TestCategory("Rule")]
@@ -64,7 +58,7 @@ namespace SonarAnalyzer.UnitTest.Rules
             Verifier.VerifyAnalyzer(@"TestCases\Hotspots\CookieShouldBeHttpOnly.cs",
                 new CS.CookieShouldBeHttpOnly(AnalyzerConfiguration.AlwaysEnabled),
                 MetadataReferenceFacade.SystemWeb,
-                CreateSonarProjectConfig(@"TestCases\WebConfig\CookieShouldBeHttpOnly\NonHttpOnlyCookiesConfig"));
+                TestHelper.CreateSonarProjectConfig(@"TestCases\WebConfig\CookieShouldBeHttpOnly\NonHttpOnlyCookiesConfig"));
 #else
         [TestMethod]
         [TestCategory("Rule")]
@@ -92,12 +86,5 @@ namespace SonarAnalyzer.UnitTest.Rules
             Verifier.VerifyAnalyzer(@"TestCases\Hotspots\CookieShouldBeHttpOnly_Nancy.cs",
                 new CS.CookieShouldBeHttpOnly(AnalyzerConfiguration.AlwaysEnabled),
                 NuGetMetadataReference.Nancy());
-
-        private static string CreateSonarProjectConfig(string filesToAnalyzeDirectory)
-        {
-            var sonarProjectConfigPath = Path.Combine(filesToAnalyzeDirectory, "SonarProjectConfig.xml");
-            File.WriteAllText(sonarProjectConfigPath, string.Format(ProjectConfigTemplate, filesToAnalyzeDirectory));
-            return sonarProjectConfigPath;
-        }
     }
 }
