+++ analyzers/tests/SonarAnalyzer.UnitTest/Rules/Hotspots/DisablingRequestValidationTest.cs
@@ -41,11 +41,6 @@ namespace SonarAnalyzer.UnitTest.Rules
     {
         private const string AspNetMvcVersion = "5.2.7";
         private const string WebConfig = "Web.config";
-        // for the tests that don't test the XML logic, to avoid test failures caused by unexpected issues from Web.config files
-        private const string ProjectConfigTemplate = @"
-<SonarProjectConfig xmlns=""http://www.sonarsource.com/msbuild/analyzer/2021/1"">
-    <FilesToAnalyzePath>{0}\FilesToAnalyze.txt</FilesToAnalyzePath>
-</SonarProjectConfig>";
 
         [TestMethod]
         [TestCategory("Rule")]
@@ -71,7 +66,7 @@ namespace SonarAnalyzer.UnitTest.Rules
                 SolutionBuilder.Create().AddProject(AnalyzerLanguage.CSharp).GetCompilation(),
                 new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled),
                 File.ReadAllText(Path.Combine(root, WebConfig)),
-                CreateSonarProjectConfig(root));
+                TestHelper.CreateSonarProjectConfig(root));
 
         [DataTestMethod]
         [DataRow(@"TestCases\WebConfig\DisablingRequestValidation\MultipleFiles", "SubFolder")]
@@ -93,7 +88,7 @@ namespace SonarAnalyzer.UnitTest.Rules
                     compilation,
                     new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled),
                     CompilationErrorBehavior.Default,
-                    sonarProjectConfigPath: CreateSonarProjectConfig(rootDirectory)).ToList();
+                    sonarProjectConfigPath: TestHelper.CreateSonarProjectConfig(rootDirectory)).ToList();
             }
             finally
             {
@@ -119,7 +114,7 @@ namespace SonarAnalyzer.UnitTest.Rules
                 SolutionBuilder.Create().AddProject(AnalyzerLanguage.CSharp).GetCompilation(),
                 new CS.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled),
                 File.ReadAllText(Path.Combine(root, "web.config")),
-                CreateSonarProjectConfig(root));
+                TestHelper.CreateSonarProjectConfig(root));
         }
 
         [DataTestMethod]
@@ -157,7 +152,7 @@ namespace SonarAnalyzer.UnitTest.Rules
                 SolutionBuilder.Create().AddProject(AnalyzerLanguage.VisualBasic).GetCompilation(),
                 new VB.DisablingRequestValidation(AnalyzerConfiguration.AlwaysEnabled),
                 File.ReadAllText(Path.Combine(root, WebConfig)),
-                CreateSonarProjectConfig(root));
+                TestHelper.CreateSonarProjectConfig(root));
         }
 
         // Verifies the results for the given web.config file path.
@@ -166,13 +161,6 @@ namespace SonarAnalyzer.UnitTest.Rules
             var actualIssues = allDiagnostics.Where(d => d.Location.GetLineSpan().Path.EndsWith(webConfigPath));
             var expectedIssues = new IssueLocationCollector().GetExpectedIssueLocations(SourceText.From(File.ReadAllText(webConfigPath)).Lines).ToList();
             DiagnosticVerifier.CompareActualToExpected(languageVersion, actualIssues, expectedIssues, false);
-        }
-
-        private static string CreateSonarProjectConfig(string filesToAnalyzeDirectory)
-        {
-            var sonarProjectConfigPath = Path.Combine(filesToAnalyzeDirectory, "SonarProjectConfig.xml");
-            File.WriteAllText(sonarProjectConfigPath, string.Format(ProjectConfigTemplate, filesToAnalyzeDirectory));
-            return sonarProjectConfigPath;
         }
     }
 }
